Add spec for Stats class, fix rare bug when calculating sequence numbers.
The sequence number for strings consisting of 2 or more spaces ("  ", "   ", …) was wrongly calculated as 0.

The botascii_tweets.txt has been slightly cleaned up to account for some bugs from the early development of BotAscii, as well as for some Twitter intricacies (eg. a tweet “m” just won't show up in your timeline).